moved generic report writer to audit helper class
